Add checking for acl constraints.
This code in its first iteration checks that the admins group
cannot be removed from the grant ACE of any object that it already
exists on. This is to protect users from getting themselves into
sticky situations, such as not being able to take admin actions
in the system.

Note that one side effect of this system is that if the admins group
is added to the grant ace of an object it was not previously on
this code will make it so it can't be removed.